Made improvements to get body mass index endpoint

+++ src/main/java/com/FitnessApp/api/model/BodyMassIndexResponse.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Builder
 public class BodyMassIndexResponse {
 
-    private Integer bodyMassIndex;
+    private Double bodyMassIndex;
 }
